server/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "server"             
version = "0.1.0"         
description = "Plataforma para venta de Repositorios."
authors = [
  { name = "Isaías Méndez", email = "encryptumv@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.9"

dependencies = [
"alembic==1.16.1",
"annotated-types==0.7.0",
"anyio==4.9.0",
"asttokens==3.0.0",
"Authlib==1.5.2",
"certifi==2025.4.26",
"cffi==1.17.1",
"charset-normalizer==3.4.2",
"click==8.2.1",
"colorama==0.4.6",
"cryptography==45.0.2",
"ecdsa==0.19.1",
"executing==2.2.0",
"fastapi==0.115.12",
"greenlet==3.2.2",
"h11==0.16.0",
"httpcore==1.0.9",
"httpx==0.28.1",
"icecream==2.1.4",
"idna==3.10",
"iniconfig==2.1.0",
"itsdangerous==2.2.0",
"Mako==1.3.10",
"MarkupSafe==3.0.2",
"packaging==25.0",
"pluggy==1.6.0",
"pyasn1==0.4.8",
"pycparser==2.22",
"pydantic==2.11.5",
"pydantic-settings==2.9.1",
"pydantic_core==2.33.2",
"Pygments==2.19.1",
"pytest==8.3.5",
"python-dotenv==1.1.0",
"python-jose==3.4.0",
"python-multipart==0.0.20",
"requests==2.32.3",
"rsa==4.9.1",
"six==1.17.0",
"sniffio==1.3.1",
"SQLAlchemy==2.0.41",
"starlette==0.46.2",
"typing-inspection==0.4.1",
"typing_extensions==4.13.2",
"urllib3==2.4.0",
"uvicorn==0.34.2",
"psycopg2-binary==2.9.7"
]

[tool.setuptools.packages.find]
where = ["src"]
